+++ projects/vdk-plugins/vdk-jupyter/vdk-jupyterlab-extension/src/__tests__/show-dialog.spec.tsx
@@ -7,9 +7,6 @@ import { jobRequest, jobRunRequest } from '../serverRequests';
 import DownloadJobDialog, {
   showDownloadJobDialog
 } from '../components/DownloadJob';
-import DeployJobDialog, {
-  showCreateDeploymentDialog
-} from '../components/DeployJob';
 import CreateJobDialog, { showCreateJobDialog } from '../components/CreateJob';
 import { VdkErrorMessage } from '../components/VdkErrorMessage';
 
@@ -132,40 +129,6 @@ describe('showDownloadJobDialog', () => {
 
     // Expect the jobRequest function to have been called with the correct arguments
     expect(jobRequest).toHaveBeenCalledWith('download');
-  });
-});
-
-describe('showCreateDeploymentDialog', () => {
-  beforeEach(() => {
-    jobData.set(VdkOption.PATH, 'test/path');
-    jobData.set(VdkOption.NAME, 'test-name');
-    jobData.set(VdkOption.TEAM, 'test-team');
-    // Mock the result of the showDialog function
-    const mockResult = { button: { accept: true }, value: true };
-    (showDialog as jest.Mock).mockResolvedValueOnce(mockResult);
-    // Mock the jobRunRequest function
-    (jobRunRequest as jest.Mock).mockResolvedValueOnce({
-      message: 'Job completed successfully!',
-      status: true
-    });
-  });
-
-  it('should show a dialog with the Create Deployment title and a DeployJobDialog component as its body', async () => {
-    // Call the function
-    await showCreateDeploymentDialog();
-
-    // Expect the first showDialog function to have been called with the correct parameters
-    expect(showDialog).toHaveBeenCalledWith({
-      title: 'Create Deployment',
-      body: (
-        <DeployJobDialog
-          jobName={jobData.get(VdkOption.NAME)!}
-          jobPath={jobData.get(VdkOption.PATH)!}
-          jobTeam={jobData.get(VdkOption.TEAM)!}
-        />
-      ),
-      buttons: [Dialog.okButton(), Dialog.cancelButton()]
-    });
   });
 });
